Refactor : Refactoring kakao page
